Optimize pipeline performance: 28% faster full pipeline (730µs → 528µs)
Five targeted optimizations across the compiler pipeline:

- Lexer cursor: byte-level ASCII fast-path for peek/advance/eat_while,
  avoiding UTF-8 iterator creation on every call (-34% lexer time)
- Parser newline suppression: O(n) swap-based compaction replaces O(n²)
  repeated Vec::remove calls (-14% parse time)
- Name type: String → Arc<str> makes clone a ref-count bump instead of
  heap allocation, benefiting all 395 clone sites across 50 files
  (-46% lower, -31% check)
- Span fields: usize → u32 shrinks every Token and AST node by 12 bytes,
  improving cache density across the entire pipeline
- Token Vec pre-allocation based on source length heuristic

+++ crates/ttrpg_ast/src/diagnostic.rs
@@ -84,8 +84,8 @@ fn render_diagnostic(
     source: &str,
     line_starts: &[usize],
 ) -> String {
-    let local_start = diag.span.start;
-    let local_end = diag.span.end;
+    let local_start = diag.span.start as usize;
+    let local_end = diag.span.end as usize;
 
     let line = line_starts
         .partition_point(|&start| start <= local_start)
